feat: new molecules
